// projects/automation/src/eventing/interfaces.ts
// Copyright (c) Perpetual-Motion project
// Licensed under the MIT License.

/* eslint-disable @typescript-eslint/naming-convention */

export interface EventData<TInput = any|undefined> {
  readonly name: string;
  completed?: Promise<unknown>;
  readonly source?: ArbitraryObject;
  readonly data: TInput;
  readonly text: string;
}

export interface Descriptor {
  name: string;
  values?: Array<string>;
}

export type Unsubscribe = () => void;

export type Callback<TOutput = unknown > = (event: EventData<TOutput>,...args:Array<any>)=>Promise<TOutput>|Promise<EventStatus>|Promise<undefined>|TOutput|EventStatus|undefined|void|Promise<void>;

export type PickByType<T, TKeepType> = {
  [P in keyof T as T[P] extends TKeepType ? P : never]: T[P]
}

export type Subscription<T extends Record<string,any> = Record<string,any>> = PickByType<T,Callback|string>;

export type ArbitraryObject = Record<string, any>;

export type EventStatus = 'Cancelled'|undefined;

export const Cancelled = 'Cancelled';
export const Continue = undefined;

export type Filter = ($data:ArbitraryObject, $strings: Array<string>, $captures: Array<string>)=> boolean;

export interface Subscriber {
  /** the filter checks generated from the event registration  */
  filters: Map<string,Filter|true>;

  /** the source (aka 'this') value that must match if the handler is tied to an object  */
  eventSource?: WeakRef<ArbitraryObject>;

  /** the event handler function itself */
  handler: Callback<any>;
}